requirements.txt
flask
google-cloud-storage
google-generativeai
python-dotenv
Werkzeug
google-ai-generativelanguage
openai